task-service/pyproject.toml
[project]
name = "task-service"
version = "0.1.0"
description = "Platform-agnostic task management service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "services"]
